+++ cli/tests/pcluster/config/test_cluster_config.py
@@ -760,7 +760,7 @@ class TestBaseClusterConfig:
             instance_type="t3.xlarge",
             image=LoginNodesImage(custom_ami="ami-0222222222222222"),
             networking=LoginNodesNetworking(subnet_ids=["subnet-0222222222222222"]),
-            ssh=LoginNodesSsh(key_name="mykey"),
+            ssh=LoginNodesSsh(),
         )
 
         login_nodes = LoginNodes(pools=[login_node_pool])
@@ -804,31 +804,31 @@ class TestBaseClusterConfig:
         assert_that(queue.job_exclusive_allocation).is_equal_to(expected_value)
 
     @pytest.mark.parametrize(
-        "head_node_ssh, login_node_ssh, expected_values",
+        "head_node_ssh, login_node_ssh, expected_allowed_ips",
         [
             (
                 HeadNodeSsh(key_name="head-node-key", allowed_ips="1.2.3.4/24"),
-                LoginNodesSsh(key_name="login-node-key", allowed_ips="6.5.4.3/24"),
-                {"key_name": "login-node-key", "allowed_ips": "6.5.4.3/24"},
+                LoginNodesSsh(allowed_ips="6.5.4.3/24"),
+                "6.5.4.3/24",
             ),
             (
                 HeadNodeSsh(key_name="head-node-key", allowed_ips="1.2.3.4/24"),
-                LoginNodesSsh(key_name=None),
-                {"key_name": "head-node-key", "allowed_ips": "1.2.3.4/24"},
+                LoginNodesSsh(),
+                "1.2.3.4/24",
             ),
             (
                 HeadNodeSsh(key_name="head-node-key", allowed_ips="1.2.3.4/24"),
                 None,
-                {"key_name": "head-node-key", "allowed_ips": "1.2.3.4/24"},
+                "1.2.3.4/24",
             ),
             (
                 None,
-                LoginNodesSsh(key_name="login-node-key"),
-                {"key_name": "login-node-key", "allowed_ips": "0.0.0.0/0"},
+                LoginNodesSsh(allowed_ips="6.5.4.3/24"),
+                "6.5.4.3/24",
             ),
         ],
     )
-    def test_login_nodes_ssh_default_values(self, head_node_ssh, login_node_ssh, expected_values, mocker):
+    def test_login_nodes_ssh_default_values(self, head_node_ssh, login_node_ssh, expected_allowed_ips, mocker):
         cluster_config = SlurmClusterConfig(
             cluster_name="clustername",
             login_nodes=LoginNodes(
@@ -863,9 +863,7 @@ class TestBaseClusterConfig:
             ),
         )
         print(cluster_config.head_node.ssh.key_name)
-        print(cluster_config.login_nodes.pools[0].ssh.key_name)
-        assert_that(cluster_config.login_nodes.pools[0].ssh.key_name).is_equal_to(expected_values["key_name"])
-        assert_that(cluster_config.login_nodes.pools[0].ssh.allowed_ips).is_equal_to(expected_values["allowed_ips"])
+        assert_that(cluster_config.login_nodes.pools[0].ssh.allowed_ips).is_equal_to(expected_allowed_ips)
 
 
 class TestSharedEbs: